chore: removed floating cta
